# Conflicts:
#	api/src/org/labkey/api/data/ColumnRenderPropertiesImpl.java
#	api/src/org/labkey/api/data/DataColumn.java
#	api/src/org/labkey/api/data/DisplayColumn.java
#	api/src/org/labkey/api/ontology/KindOfQuantity.java
#	api/src/org/labkey/api/ontology/Quantity.java
#	api/src/org/labkey/api/ontology/Unit.java
#	experiment/resources/schemas/dbscripts/postgresql/exp-25.005-25.006.sql
#	experiment/resources/schemas/dbscripts/sqlserver/exp-25.005-25.006.sql

+++ api/src/org/labkey/api/di/DataIntegrationService.java
@@ -59,7 +59,7 @@ public interface DataIntegrationService
     /**
      * Returns the Set of transform job IDs for the container
      */
-    Set<Integer> getTransformRunJobIds(Container c);
+    Set<Long> getTransformRunJobIds(Container c);
 
     /**
      * Execute an efficient reimport operation Create a dataIterator based on target.getQueryUpdateService()
